refactor(document processor): Move extract_from_pdf to lib
