+++ cl/lib/document_processors.py
@@ -5,6 +5,7 @@ from PyPDF2 import PdfFileReader
 from PyPDF2.utils import PdfReadError
 
 from cl.celery import app
+from cl.lib.mojibake import fix_mojibake
 from cl.scrapers.tasks import DEVNULL, rasterize_pdf, convert_file_to_txt, \
     cleanup_ocr_text
 
@@ -62,3 +63,26 @@ def extract_by_ocr(path):
         txt = cleanup_ocr_text(txt)
 
     return True, txt
+
+
+def extract_from_pdf(path, opinion, do_ocr=False):
+    """ Extract text from pdfs.
+
+    Here, we use pdftotext. If that fails, try to use tesseract under the
+    assumption it's an image-based PDF. Once that is complete, we check for the
+    letter e in our content. If it's not there, we try to fix the mojibake
+    that ca9 sometimes creates.
+    """
+    process = make_pdftotext_process(path)
+    content, err = process.communicate()
+    if content.strip() == '' and do_ocr:
+        success, content = extract_by_ocr(path)
+        if success:
+            opinion.extracted_by_ocr = True
+        elif content == '' or not success:
+            content = 'Unable to extract document content.'
+    elif 'e' not in content:
+        # It's a corrupt PDF from ca9. Fix it.
+        content = fix_mojibake(unicode(content, 'utf-8', errors='ignore'))
+
+    return content, err
